Rename executeInline to execute and executeStream to executeAndStreamResults.
The names of the current methods were misleading and did not clarify the
intent of the methods. The executeInline method loaded and returned all
the results at once, while the executeStream method returned a Readble
which then streamed one result at a time(while managing pagination in
background).
The new method names aim to disamiguate the functionality of the two
methods