Simplify session and executable interfaces
We had QldbSession interface which was implemented by the
QldbSessionImpl and PooledQldbSesssion. The motivation was to support
the sessions for pooled and non-pooled driver. But with the removal of the
driver, this is no more required. Therefore, in this commit I have removed
the interface QldbSession and renamed QldbSessionImpl.
Also, the executable interface did not make sense anymore.

So there will now be only one class for the driver and one class for the
session