Expose only the executeLambda method of driver
in the earlier versions of the driver, the customers could execute the
Transaction by calling executeLambda method on the pooledQLdb driver
instance or on the session instance. This lead to confusion and there
was no documentation which helped customers choose one over the other.
Moreover, there was no performance benefit of doing one over the other.

This change fixes couple of things. First, it will expose only
executeLambda method to the customers thereby allowing the customers to
run their Tx only in one way.
Secondly, it fixes the pooling logic. Earlier, while executing the Tx,
if there was an invalidSessionException thrown from the backend, the
driver would send another request to create a new session and retry the
Tx with the new session. This mechanism did not take advantage of a
possibly healthy session in the pool and maybe we could have avoided
creating new session. This change fixes that and now, whenever ther eis
ISE, the driver will try with the next session in the pool. If all the
sessions in the pool are exhausted, only then the driver will create a
new session