Abort on LambdaAbortedException
When the customer calls abort on TransactionExecutor, the driver just
throws the exception. But we have to actually send the abort as well
since it might leave a dangling transaction.

This commit removes the check for LambdaAbortedException. With that in
place the abort will be done via the generic condition(abort if not OCC)
and then the error will be thrown back